requirements-dev.txt
tensorflow
keras
pillow
numpy
scikit-learn
flask
flask-cors
pandas
seaborn
matplotlib
openpyxl
